removed cn class, bugfixes for jupyter tutorial
