+++ autocnet/control/control.py
@@ -120,126 +120,3 @@ def identify_potential_overlaps(cg, cn, overlap=True):
 
 def deepen_correspondences(cg, cn):
     pass
-
-class ControlNetwork(object):
-    measures_keys = ['point_id', 'image_index', 'keypoint_index', 'edge', 'match_idx', 'x', 'y', 'x_off', 'y_off', 'corr', 'valid']
-
-    def __init__(self):
-        self._point_id = 0
-        self._measure_id = 0
-        self.measure_to_point = {}
-        self.data = pd.DataFrame(columns=self.measures_keys)
-
-    @classmethod
-    def from_candidategraph(cls, matches):
-        cls = ControlNetwork()
-        for match in matches:
-            for idx, row in match.iterrows():
-                edge = (row.source_image, row.destination_image)
-                source_key = (row.source_image, row.source_idx)
-                source_fields = row[['source_x', 'source_y']]
-                destin_key = (row.destination_image, row.destination_idx)
-                destin_fields = row[['destination_x', 'destination_y']]
-                if cls.measure_to_point.get(source_key, None) is not None:
-                    tempid = cls.measure_to_point[source_key]
-                    cls.add_measure(destin_key, edge, row.name, destin_fields, point_id=tempid)
-                elif cls.measure_to_point.get(destin_key, None) is not None:
-                    tempid = cls.measure_to_point[destin_key]
-                    cls.add_measure(source_key, edge, row.name,  source_fields, point_id=tempid)
-                else:
-                    cls.add_measure(source_key, edge, row.name,  source_fields)
-                    cls.add_measure(destin_key, edge,row.name,  destin_fields)
-                    cls._point_id += 1
-
-        cls.data.index.name = 'measure_id'
-        return cls
-
-    def add_measure(self, key, edge, match_idx, fields, point_id=None):
-        """
-        Create a new measure that is coincident to a given point.  This method does not
-        create the point if is missing.  When a measure is added to the graph, an associated
-        row is added to the measures dataframe.
-
-        Parameters
-        ----------
-        key : hashable
-                  Some hashable id.  In the case of an autocnet graph object the
-                  id should be in the form (image_id, match_id)
-
-        point_id : hashable
-                   The point to link the node to.  This is most likely an integer, but
-                   any hashable should work.
-        """
-        if key in self.measure_to_point.keys():
-            return
-        if point_id == None:
-            point_id = self._point_id
-        self.measure_to_point[key] = point_id
-        # The node_id is a composite key (image_id, correspondence_id), so just grab the image
-        image_id = key[0]
-        match_id = key[1]
-        self.data.loc[self._measure_id] = [point_id, image_id, match_id, edge, match_idx, *fields, 0, 0, np.inf, True]
-        self._measure_id += 1
-
-    def remove_measure(self, idx):
-        self.data = self.data.drop(self.data.index[idx])
-        for r in idx:
-            self.measure_to_point.pop(r, None)
-
-    def validate_points(self):
-        """
-        Ensure that all control points currently in the nework are valid.
-
-        Criteria for validity:
-
-          * Singularity: A control point can have one and only one measure from any image
-
-        Returns
-        -------
-         : pd.Series
-
-        """
-
-        def func(g):
-            # One and only one measure constraint
-            if g.image_index.duplicated().any():
-                return True
-            else: return False
-        return self.data.groupby('point_id').apply(func)
-
-    def clean_singles(self):
-        """
-        Take the `data` dataframe and return only those points with
-        at least two measures.  This is automatically called before writing
-        as functions such as subpixel matching can result in orphaned measures.
-        """
-        return self.data.groupby('point_id').apply(lambda g: g if len(g) > 1 else None)
-
-    def to_isis(self, outname, serials, olist, *args, **kwargs): #pragma: no cover
-        """
-        Write the control network out to the ISIS3 control network format.
-        """
-
-        if self.validate_points().any() == True:
-            warnings.warn('Control Network is not ISIS3 compliant.  Please run the validate_points method on the control network.')
-            return
-
-        # Apply the subpixel shift
-        self.data.x += self.data.x_off
-        self.data.y += self.data.y_off
-
-        to_isis(outname + '.net', self.data.query('valid == True'),
-                serials, *args, **kwargs)
-        write_filelist(olist, outname + '.lis')
-
-        # Back out the subpixel shift
-        self.data.x -= self.data.x_off
-        self.data.y -= self.data.y_off
-
-    def to_bal(self):
-        """
-        Write the control network out to the Bundle Adjustment in the Large
-        (BAL) file format.  For more information see:
-        http://grail.cs.washington.edu/projects/bal/
-        """
-        pass
